Watch now responds to files that are replaced
When you watch a file you are watching the file not the path,
so when it's removed and replace your watcher is basically dead.

We changed the watcher to now point to the parent directory and
check WRITE events to see if they match any of the watches

It would probably be more efficient to have a single watcher
for everything, not separate watchers for each process.